requirements.txt
# SECS Driver 依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
PyYAML>=6.0
